new version to replace pyswagger.spec.base.BaseObj
to fix the mess introduced by ContainerType, in old way, to support
objects organized by combination of list and dict is just not easy. (or
field with different types of possible value, ex AdditionalProperties

new way is base2.Base2Obj, and support “function as class constructor”
when constructing the hierarchy when traversing spec.